% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_account_dca.R
\name{get_account_dca}
\alias{get_account_dca}
\title{Get Accounts of Annual Reports}
\usage{
get_account_dca(year, annex, entity, grouped_by_year = FALSE,
  grouped_by_entity = FALSE)
}
\arguments{
\item{year}{a numeric vector}

\item{annex}{character}

\item{entity}{a character vector. IBGE codes of a public sector entities set.}

\item{grouped_by_year}{boolean. Default = FALSE}

\item{grouped_by_entity}{boolean. Default = FALSE}
}
\value{
dataframe
}
\description{
Get Accounts of Annual Reports
}
\examples{
get_account_dca(c(2018), "I-E", c("2402600") )
}
