% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dca.R
\name{get_dca}
\alias{get_dca}
\title{Get Annual Accounts Data}
\usage{
get_dca(year, annex, entity, arg_cod_conta = NULL, In_QDCC = FALSE)
}
\arguments{
\item{year}{a numeric vector}

\item{annex}{character}

\item{entity}{a character vector. IBGE code of a public sector entities set.}

\item{arg_cod_conta}{a character vector. Set of a accounts codes to filter the dataset. Use get_account_dca to see the accounts available. Default = NULL}

\item{In_QDCC}{boolean. Default = FALSE}
}
\value{
dataframe
}
\description{
Get data related to annual accounts reports of a given set of entities of the public brazilian sector, filtered by a set of parameters passed to the function.
For more information about these reports, please see https://siconfi.tesouro.gov.br/siconfi/pages/public/conteudo/conteudo.jsf?id=82
}
\examples{
get_dca(year = c(2018), annex = "I-C",  entity = "21")
}
