% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dca_mun_state.R
\name{get_dca_mun_state}
\alias{get_dca_mun_state}
\title{Get Municipalities Annual Accounts Data}
\usage{
get_dca_mun_state(year, annex, state, arg_cod_conta = NULL,
  In_QDCC = FALSE)
}
\arguments{
\item{year}{a numeric vector}

\item{annex}{character}

\item{state}{state a character vector. IBGE code of a given Brazilian states set.}

\item{arg_cod_conta}{a character vector. Set of a accounts codes to filter the dataset. Use get_account_dca to see the accounts available. Default = NULL}

\item{In_QDCC}{boolean. Default = FALSe}
}
\value{
dataframe
}
\description{
Get data related to annual accounts for all the municipalities of a given Brazilian states set, filtered by a set of parameters passed to the function.
For more information about these reports, please see https://siconfi.tesouro.gov.br/siconfi/pages/public/conteudo/conteudo.jsf?id=82
}
\examples{
get_dca_mun_state( year = c(2018),  annex = "I-C",  arg_cod_conta = c("TotalReceitas", "1.7.0.0.00.0.0", "2.4.0.0.00.0.0"),  state = c(21:29))
}
