% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rgf.R
\name{get_rgf}
\alias{get_rgf}
\title{Get Fiscal Managment Data}
\usage{
get_rgf(year, periodicity, period, report_tp, annex, entity, co_power)
}
\arguments{
\item{year}{a numeric vector}

\item{periodicity}{character. Available values: Q, S}

\item{period}{a numeric vector. Available values: 1, 2, 3}

\item{report_tp}{numeric. Available values: 1- RGF, 2-RGF Simplificado}

\item{annex}{character}

\item{entity}{a character vector. IBGE code of a public sector entities set.}

\item{co_power}{a character vector.Available values:  E = Executivo, L = Legislativo, J = Judiciário, M = Ministério Público, D = Defensoria Pública}
}
\value{
dataframe
}
\description{
Get data related to fiscal management reports of a given set of entities of the public brazilian sector, filtered by a set of parameters passed to the function.
For more information about these reports, please see https://siconfi.tesouro.gov.br/siconfi/pages/public/conteudo/conteudo.jsf?id=82
}
\examples{
get_rgf(year=  c(2018),  periodicity = "Q",  period =  1,    report_tp = 1,      annex = "02",       entity = c("2312908"),   co_power = "E")
}
