% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rreo.R
\name{get_rreo}
\alias{get_rreo}
\title{Get Budgetary Execution Report Data}
\usage{
get_rreo(year, period, report_tp, annex, entity, In_RGPS = FALSE,
  In_RPPS = FALSE)
}
\arguments{
\item{year}{a numeric vector}

\item{period}{a numeric vector. Available values: 1, 2, 3, 4, 5, 6}

\item{report_tp}{numeric. Available values: 1- RREO, 2-RREO Simplificado}

\item{annex}{character}

\item{entity}{a character vector. IBGE code of a public sector entities set.}

\item{In_RGPS}{boolean. Default = FALSE}

\item{In_RPPS}{boolean. Default = FALSE}
}
\value{
dataframe
}
\description{
Get data related to budgetary execution reports of a given set of entities of the public brazilian sector, filtered by a set of parameters passed to the function.
For more information about these reports, please see https://siconfi.tesouro.gov.br/siconfi/pages/public/conteudo/conteudo.jsf?id=82
}
\examples{
get_rreo(2017,1,1,"01","1")
}
