% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rreo_mun_state.R
\name{get_rreo_mun_state}
\alias{get_rreo_mun_state}
\title{Get Municipalities Budgetary Execution Report Data}
\usage{
get_rreo_mun_state(year, period, report_tp, annex, state,
  In_RGPS = FALSE, In_RPPS = FALSE)
}
\arguments{
\item{year}{a numeric vector}

\item{period}{a numeric vector. Available values: 1, 2, 3, 4, 5, 6}

\item{report_tp}{numeric. Available values: 1- RREO, 2-RREO Simplificado}

\item{annex}{character}

\item{state}{a character vector. IBGE code of a given Brazilian states set.}

\item{In_RGPS}{boolean. Default = FALSE}

\item{In_RPPS}{boolean. Default = FALSE}
}
\value{
A dataframe with the result set of a given entity of the public brazilian set, filtered by a set of parameters passed to the function
}
\description{
#' @description
Get data related to budgetary execution reports for all the municipalities of a given Brazilian states set, filtered by a set of parameters passed to the function.
For more information about these reports, please see https://siconfi.tesouro.gov.br/siconfi/pages/public/conteudo/conteudo.jsf?id=82
}
\examples{
get_rreo_mun_state(2017,1,1,"01","11")
}
