% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rgf_mun_state.R
\name{get_rgf_mun_state}
\alias{get_rgf_mun_state}
\title{Get Municipalities Fiscal Managment Data}
\usage{
get_rgf_mun_state(year, periodicity, period, report_tp, annex, state,
  co_power)
}
\arguments{
\item{year}{a numeric vector}

\item{periodicity}{character. Available values: Q, S}

\item{period}{a numeric vector. Available values: 1, 2, 3}

\item{report_tp}{numeric. Available values: 1- RGF, 2-RGF Simplificado}

\item{annex}{character}

\item{state}{state a character vector. IBGE code of a given Brazilian states set.}

\item{co_power}{a character vector.Available values:  E = Executivo, L = Legislativo, J = Judiciário, M = Ministério Público, D = Defensoria Pública}
}
\value{
dataframe
}
\description{
Get Municipalities Fiscal Managment Data
}
\examples{
get_rgf_mun_state(year=  c(2018),  periodicity = "Q",  period =  3, report_tp = 1,  annex = "01",  state = c(11:17), co_power = "E")
}
